IEP-1089: Updated the project creation wizard to have only one page with templates as well
