+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/templates/messages.properties
@@ -1,6 +1,7 @@
 BaseWizardSelectionPage_No_Desc=No Description available.
 TemplateListSelectionPage_SelectTemplate_Desc=&Create a project using one of the templates
 TemplateListSelectionPage_Template_Wizard_Desc=Select one of the available templates to generate a fully-functioning IDF project.
-TemplateListSelectionPage_Templates=Templates
-TemplateListSelectionPage_Templates_Desc=Select one of the available templates to generate a fully-functioning IDF project.
+NewProjectWizardPage_Header=New Project
+NewProjectWizardPage_DescriptionString=Select one of the available templates to generate a fully-functioning IDF project or create your own
 TemplateSelectionToolTip=To choose a template, select the "Create a project using one of the templates" checkbox
+NewProjectWizardPage_NoTemplateFoundMessage=No Templates were found on the system you can continue to create a blank project
